+++ board-frontend/html/scripts/protocol/server/messages/ServerMessageDeleteDocument.js
@@ -1,0 +1,9 @@
+import ServerMessage from "../ServerMessage.js";
+import ServerMessageType from "../ServerMessageType.js";
+
+export default class ServerMessageDeleteDocument extends ServerMessage {
+    constructor(reader) {
+        super(ServerMessageType.DELETE_DOCUMENT);
+        this.id = reader.readBigInt64();
+    }
+}