+++ board-room/src/main/java/net/stzups/board/BoardRoom.java
@@ -1,18 +1,18 @@
 package net.stzups.board;
 
 import io.netty.channel.ChannelFuture;
-import net.stzups.board.config.Config;
-import net.stzups.board.config.ConfigBuilder;
-import net.stzups.board.config.configs.ArgumentConfig;
-import net.stzups.board.config.configs.EnvironmentVariableConfig;
-import net.stzups.board.config.configs.PropertiesConfig;
+import net.stzups.board.util.config.Config;
+import net.stzups.board.util.config.ConfigBuilder;
+import net.stzups.board.util.LogFactory;
+import net.stzups.board.util.config.configs.ArgumentConfig;
+import net.stzups.board.util.config.configs.EnvironmentVariableConfig;
+import net.stzups.board.util.config.configs.PropertiesConfig;
 import net.stzups.board.data.database.Database;
-import net.stzups.board.data.database.postgres.PostgresDatabase;
 import net.stzups.board.data.database.memory.MemoryDatabase;
+import net.stzups.board.data.database.postgres.PostgresDatabase;
 import net.stzups.board.server.Server;
 
 import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.util.Random;
 import java.util.logging.Logger;
@@ -20,9 +20,7 @@ import java.util.logging.Logger;
 public class BoardRoom {
     private static Logger logger;
     private static Config config;
-    private static final SecureRandom secureRandom = new SecureRandom();
     private static final Random random = new Random();
-    private static MessageDigest SHA256MessageDigest;
 
     private static Database database;//user id -> user
 
@@ -33,26 +31,21 @@ public class BoardRoom {
 
         long start = System.currentTimeMillis();
 
-        SHA256MessageDigest = MessageDigest.getInstance("SHA-256");
-
         config = new ConfigBuilder()
                 .addConfig(new ArgumentConfig(args))
                 .addConfig(new PropertiesConfig("board.properties"))
                 .addConfig(new EnvironmentVariableConfig("board."))
                 .build();
 
-        Boolean postgres = config.getBoolean("postgres");
-        if (postgres == null) {
-            throw new RuntimeException("Failed to specify required runtime variable --postgres");
+        if (BoardRoom.getConfig().getBoolean(BoardConfigKeys.POSTGRES)) {
+            logger.info("Connecting to Postgres database...");
+            database = new PostgresDatabase(BoardRoom.getConfig().getString(BoardConfigKeys.POSTGRES_URL),
+                    BoardRoom.getConfig().getString(BoardConfigKeys.POSTGRES_USER),
+                    BoardRoom.getConfig().getString(BoardConfigKeys.POSTGRES_PASSWORD));
+            logger.info("Connected to Postgres database");
         } else {
-            if (postgres) {
-                logger.info("Connecting to Postgres database...");
-                database = new PostgresDatabase(BoardRoom.getConfig().get("postgres.url"), BoardRoom.getConfig().get("postgres.user"), BoardRoom.getConfig().get("postgres.password"));
-                logger.info("Connected to Postgres database");
-            } else {
-                logger.warning("Using debug only runtime database. No data will be persisted.");
-                database = new MemoryDatabase();
-            }
+            logger.warning("Using debug only runtime database. No data will be persisted.");
+            database = new MemoryDatabase();
         }
 
         Server server = new Server();
@@ -81,14 +74,6 @@ public class BoardRoom {
 
     public static Database getDatabase() {
         return database;
-    }
-
-    public static SecureRandom getSecureRandom() {
-        return secureRandom;
-    }
-
-    public static MessageDigest getSHA256MessageDigest() {
-        return SHA256MessageDigest;
     }
 
     public static Random getRandom() {