+++ board-room/src/main/java/net/stzups/board/config/ConfigProvider.java
@@ -1,5 +1,0 @@
-package net.stzups.board.config;
-
-public interface ConfigProvider {
-    String get(String key);
-}