+++ board-room/src/main/java/net/stzups/board/data/database/memory/MemoryDatabase.java
@@ -1,11 +1,14 @@
 package net.stzups.board.data.database.memory;
 
+import net.stzups.board.BoardRoom;
 import net.stzups.board.data.database.Database;
 import net.stzups.board.data.objects.Document;
+import net.stzups.board.data.objects.InviteCode;
 import net.stzups.board.data.objects.User;
 import net.stzups.board.data.objects.PersistentUserSession;
 import net.stzups.board.data.objects.canvas.Canvas;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -18,13 +21,20 @@ public class MemoryDatabase implements Database {
     private Map<Long, User> users = new HashMap<>();
 
     @Override
-    public void addUser(User user) {
+    public User createUser() {
+        User user = new User(BoardRoom.getRandom().nextLong(), new Long[0], new Long[0]);
         users.put(user.getId(), user);
+        return user;
     }
 
     @Override
     public User getUser(long id) {
         return users.get(id);
+    }
+
+    @Override
+    public void updateUser(User user) {
+
     }
 
     @Override
@@ -43,8 +53,23 @@ public class MemoryDatabase implements Database {
     }
 
     @Override
-    public void saveDocument(Document document) {
+    public InviteCode getInviteCode(String code) {
+        return null;
+    }
+
+    @Override
+    public InviteCode getInviteCode(Document document) {
+        return null;
+    }
+
+    @Override
+    public void updateDocument(Document document) {
         documents.put(document.getId(), document);
+    }
+
+    @Override
+    public void deleteDocument(Document document) {
+        documents.remove(document.getId());
     }
 
     @Override