+++ board-room/src/main/java/net/stzups/board/data/objects/canvas/Canvas.java
@@ -1,6 +1,7 @@
 package net.stzups.board.data.objects.canvas;
 
 import io.netty.buffer.ByteBuf;
+import net.stzups.board.BoardRoom;
 import net.stzups.board.data.objects.Document;
 import net.stzups.board.data.objects.canvas.object.CanvasObject;
 import net.stzups.board.data.objects.canvas.object.CanvasObjectType;
@@ -23,11 +24,13 @@ public class Canvas {
      */
     public Canvas(Document document, ByteBuf byteBuf) {
         this.document = document;
-        for (int i = 0; i < byteBuf.readUnsignedByte(); i++) {
+        int length = byteBuf.readUnsignedByte();
+        for (int i = 0; i < length; i++) {
             CanvasObjectType canvasObjectType = CanvasObjectType.valueOf(byteBuf.readUnsignedByte());
             Map<Short, CanvasObject> map = new HashMap<>();
             canvasObjects.put(canvasObjectType, map);
-            for (int j = 0; j < byteBuf.readUnsignedShort(); j++) {
+            int l = byteBuf.readUnsignedShort();
+            for (int j = 0; j < l; j++) {
                 map.put(byteBuf.readShort(), CanvasObject.getCanvasObject(canvasObjectType, byteBuf));
             }
         }
@@ -38,7 +41,6 @@ public class Canvas {
     }
 
     public void update(Map<CanvasObjectType, Map<Short, CanvasObjectWrapper>> updateCanvasObjects) {
-        System.out.println(canvasObjects.size());
         for (Map.Entry<CanvasObjectType, Map<Short, CanvasObjectWrapper>> entry : updateCanvasObjects.entrySet()) {
             Map<Short, CanvasObject> map = canvasObjects.computeIfAbsent(entry.getKey(), k -> new HashMap<>());
             for (Map.Entry<Short, CanvasObjectWrapper> entry1 : entry.getValue().entrySet()) {
@@ -47,25 +49,17 @@ public class Canvas {
         }
     }
 
-    public void delete(Canvas canvas) {
-        delete(canvas.canvasObjects);
-    }
-
     public void delete(Map<CanvasObjectType, Map<Short, CanvasObject>> deleteCanvasObjects) {
         for (Map.Entry<CanvasObjectType, Map<Short, CanvasObject>> entry : deleteCanvasObjects.entrySet()) {
             Map<Short, CanvasObject> map = canvasObjects.get(entry.getKey());
             if (map == null) {
-                System.out.println("cant delete " + entry.getKey() + ", its already gone");
+                BoardRoom.getLogger().warning("Tried to delete canvas object that does not exist");
             } else {
                 for (Map.Entry<Short, CanvasObject> entry1 : entry.getValue().entrySet()) {
                     map.remove(entry1.getKey());
                 }
             }
         }
-    }
-
-    public void clear() {
-        canvasObjects.clear();
     }
 
     public void serialize(ByteBuf byteBuf) {
@@ -78,10 +72,6 @@ public class Canvas {
                 entry1.getValue().serialize(byteBuf);
             }
         }
-    }
-
-    public boolean isEmpty() {
-        return canvasObjects.isEmpty();
     }
 
     @Override