+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/client/messages/ClientMessageCreateDocument.java
@@ -1,10 +1,11 @@
 package net.stzups.board.server.websocket.protocol.client.messages;
 
+import io.netty.buffer.ByteBuf;
 import net.stzups.board.server.websocket.protocol.client.ClientMessage;
 import net.stzups.board.server.websocket.protocol.client.ClientMessageType;
 
 public class ClientMessageCreateDocument extends ClientMessage {
-    public ClientMessageCreateDocument() {
+    public ClientMessageCreateDocument(ByteBuf byteBuf) {
         super(ClientMessageType.CREATE_DOCUMENT);
     }
 }