+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/server/messages/ServerMessageDeleteDocument.java
@@ -1,0 +1,21 @@
+package net.stzups.board.server.websocket.protocol.server.messages;
+
+import io.netty.buffer.ByteBuf;
+import net.stzups.board.data.objects.Document;
+import net.stzups.board.server.websocket.protocol.server.ServerMessage;
+import net.stzups.board.server.websocket.protocol.server.ServerMessageType;
+
+public class ServerMessageDeleteDocument extends ServerMessage {
+    private long id;
+
+    public ServerMessageDeleteDocument(Document document) {
+        super(ServerMessageType.DELETE_DOCUMENT);
+        id = document.getId();
+    }
+
+    @Override
+    public void serialize(ByteBuf byteBuf) {
+        super.serialize(byteBuf);
+        byteBuf.writeLong(id);
+    }
+}