+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/server/messages/ServerMessageUpdateCanvas.java
@@ -1,10 +1,8 @@
 package net.stzups.board.server.websocket.protocol.server.messages;
 
 import io.netty.buffer.ByteBuf;
-import net.stzups.board.data.objects.canvas.Canvas;
 import net.stzups.board.data.objects.canvas.object.CanvasObjectType;
 import net.stzups.board.data.objects.canvas.object.CanvasObjectWrapper;
-import net.stzups.board.server.websocket.Client;
 import net.stzups.board.server.websocket.protocol.server.ServerMessage;
 import net.stzups.board.server.websocket.protocol.server.ServerMessageType;