+++ frontend/html/scripts/protocol/client/ClientMessageType.js
@@ -1,7 +1,0 @@
-const ClientMessageType = {
-    OPEN_DOCUMENT:0,
-    UPDATE_CANVAS:1,
-    CREATE_DOCUMENT:2,
-    HANDSHAKE:3,
-};
-export default ClientMessageType;